+++ modules/flowable-dmn-rest/src/main/java/org/flowable/dmn/rest/service/api/DmnRestUrls.java
@@ -106,36 +106,6 @@ public final class DmnRestUrls {
     public static final String[] URL_DEPLOYMENT_RESOURCE_CONTENT = { SEGMENT_REPOSITORY_RESOURCES, SEGMENT_DEPLOYMENT_RESOURCE, "{0}", SEGMENT_DEPLOYMENT_ARTIFACT_RESOURCE_CONTENT, "{1}" };
 
     /**
-     * URL template for a decision executor: <i>/dmn-decision/execute</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_RESOURCE};
-
-    /**
-     * URL template for a decision executor: <i>/dmn-decision/execute/single-result</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE_SINGLE_RESULT = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
-
-    /**
-     * URL template for a decision executor: <i>/dmn-decision/execute-decision</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE_DECISION = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_DECISION_RESOURCE};
-
-    /**
-     * URL template for a decision executor: <i>/dmn-decision/execute-decision/single-result</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE_DECISION_SINGLE_RESULT = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_DECISION_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
-
-    /**
-     * URL template for a decision executor: <i>/dmn-decision/execute-decision-service</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE_DECISION_SERVICE = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_DECISION_SERVICE_RESOURCE};
-
-    /**
-     * URL template for a decision executor: <i>/dmn-decision/execute-decision-service/single-result</i>
-     */
-    public static final String[] URL_DECISION_SERVICE_EXECUTE_DECISION_SERVICE_SINGLE_RESULT = { SEGMENT_DECISION_RESOURCES, SEGMENT_EXECUTE_DECISION_SERVICE_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
-
-    /**
      * URL template for a decision executor: <i>/dmn-rule/execute</i>
      */
     public static final String[] URL_RULE_SERVICE_EXECUTE = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_RESOURCE};
@@ -144,7 +114,27 @@ public final class DmnRestUrls {
      * URL template for a decision executor: <i>/dmn-rule/execute/single-result</i>
      */
     public static final String[] URL_RULE_SERVICE_EXECUTE_SINGLE_RESULT = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
-    
+
+    /**
+     * URL template for a decision executor: <i>/dmn-decision/execute-decision</i>
+     */
+    public static final String[] URL_RULE_SERVICE_EXECUTE_DECISION = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_DECISION_RESOURCE};
+
+    /**
+     * URL template for a decision executor: <i>/dmn-decision/execute-decision/single-result</i>
+     */
+    public static final String[] URL_RULE_SERVICE_EXECUTE_DECISION_SINGLE_RESULT = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_DECISION_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
+
+    /**
+     * URL template for a decision executor: <i>/dmn-decision/execute-decision-service</i>
+     */
+    public static final String[] URL_RULE_SERVICE_EXECUTE_DECISION_SERVICE = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_DECISION_SERVICE_RESOURCE};
+
+    /**
+     * URL template for a decision executor: <i>/dmn-decision/execute-decision-service/single-result</i>
+     */
+    public static final String[] URL_RULE_SERVICE_EXECUTE_DECISION_SERVICE_SINGLE_RESULT = { SEGMENT_RULES_RESOURCES, SEGMENT_EXECUTE_DECISION_SERVICE_RESOURCE, SEGMENT_EXECUTE_SINGLE_RESULT_RESOURCE};
+
     /**
      * URL template for a historic decision execution collection: <i>/dmn-history/historic-decision-executions</i>
      */
